# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.BackupErrorResponse:
    properties:
      error:
        example: failed to execute backup
        type: string
      message:
        example: connection management with id=123 not found
        type: string
    type: object
  controllers.BackupJobStartResponse:
    properties:
      job_id:
        example: backup_123
        type: string
      message:
        example: Backup job started successfully with 3 steps
        type: string
      status:
        example: success
        type: string
    type: object
  controllers.BackupValidationErrorResponse:
    properties:
      error:
        example: validation failed
        type: string
      message:
        example: 'invalid job ID: must be greater than 0'
        type: string
    type: object
  controllers.BulkCreateResponse:
    properties:
      message:
        example: Inserted 5 new records
        type: string
    type: object
...
  controllers.ConnectionMgtRequest:
    properties:
      cntmgt:
        example: 1
        type: integer
    type: object
...
  controllers.DBMgtCreateRequest:
    properties:
      cntmgt:
        example: 1
        type: integer
      dbname:
        example: test_database
        type: string
      dbtype:
        example: mysql
        type: string
      status:
        example: active
        type: string
    type: object
  controllers.DBMgtCreateResponse:
    properties:
      id:
        example: 123
        type: integer
      message:
        example: Database was created successfully
        type: string
    type: object
  controllers.DBMgtDeleteResponse:
    properties:
      message:
        example: Database was deleted successfully
        type: string
    type: object
...
  controllers.DBObjectMgtItem:
    properties:
      dbmgt:
        example: 1
        type: integer
      description:
        example: Table description
        type: string
      id:
        example: 1
        type: integer
      objectName:
        example: table1
        type: string
      objectTypeId:
        example: 1
        type: integer
      sqlParam:
        example: param1=value1
        type: string
      status:
        example: active
        type: string
    type: object
  controllers.DBObjectMgtListResponse:
    properties:
      message:
        items:
          $ref: '#/definitions/controllers.DBObjectMgtItem'
        type: array
    type: object
...
  controllers.DatabaseConnectionErrorResponse:
    properties:
      error:
        example: database connection failed
        type: string
    type: object
  controllers.DatabaseCreationErrorResponse:
    properties:
      error:
        example: failed to create database
        type: string
    type: object
...
  controllers.InternalServerErrorResponse:
    properties:
      error:
        example: database connection failed
        type: string
    type: object
...
  controllers.JobProcessingErrorResponse:
    properties:
      error:
        example: failed to start background job
        type: string
    type: object
  controllers.JobStartResponse:
    properties:
      message:
        example: 'Background job started: job_12345'
        type: string
    type: object
...
  controllers.StandardErrorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
...
  models.BackupRequest:
    properties:
      cnt_id:
        example: 456
        type: integer
      command:
        example: 7b2273746570733a205b7b226f72646572223a20312c2022636f6d6d616e64223a20226d7973716c64756d70202e2e2e227d5d7d
        type: string
      file_name:
        example: backup_20240101.sql
        type: string
      job_id:
        example: 123
        type: integer
      type:
        enum:
        - dump
        - check_binlog
        example: dump
        type: string
    required:
    - cnt_id
    - command
    - job_id
    - type
    type: object
...
info:
  contact: {}
  description: DBF Artifact API
  title: dbfartifactapi
  version: "1.0"
paths:
...
  /api/queries/backup:
    post:
      consumes:
      - application/json
      description: Starts background jobs to execute backup steps (dump, binlog check,
        etc.) via VeloArtifact. Each step is executed as a separate background job.
        For dump type with fileName, supports both OS commands (os_execute) and SQL
        commands (execute). For other types like check_binlog, all steps are SQL commands.
      parameters:
      - description: Backup request with hex-encoded command containing steps
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BackupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Background jobs started successfully with job IDs
          schema:
            $ref: '#/definitions/controllers.BackupJobStartResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/controllers.BackupValidationErrorResponse'
        "500":
          description: Internal server error during backup execution
          schema:
            $ref: '#/definitions/controllers.BackupErrorResponse'
      summary: Execute backup operation
      tags:
      - Backup
  /api/queries/connection/test/{id}:
    post:
      consumes:
      - application/json
      description: Tests connection to database and updates status in the system
      parameters:
      - description: Connection Management ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Connection test successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Connection not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Test database connection
      tags:
      - Connection
...
  /api/queries/dbmgt:
    post:
      consumes:
      - application/json
      description: Creates a new database management entry with specified parameters
      parameters:
      - description: Database Management object
        in: body
        name: database
        required: true
        schema:
          $ref: '#/definitions/controllers.DBMgtCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Database created successfully
          schema:
            $ref: '#/definitions/controllers.DBMgtCreateResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.DatabaseCreationErrorResponse'
      summary: Create database management
      tags:
      - DB Management
  /api/queries/dbmgt/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing database management entry by ID
      parameters:
      - description: Database Management ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Database deleted successfully
          schema:
            $ref: '#/definitions/controllers.DBMgtDeleteResponse'
        "400":
          description: Invalid database management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.InternalServerErrorResponse'
      summary: Delete database management
      tags:
      - DB Management
  /api/queries/dbmgt/all:
    post:
      consumes:
      - application/json
      description: Creates all databases for the specified connection management ID
      parameters:
      - description: Connection Management ID parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/controllers.ConnectionMgtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Databases created successfully with count
          schema:
            $ref: '#/definitions/controllers.BulkCreateResponse'
        "400":
          description: Invalid connection management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.DatabaseCreationErrorResponse'
      summary: Create all databases for connection management
      tags:
      - DB Management
...
  /api/queries/dbobjectmgt/dbmgt/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all database objects associated with the specified database
        management ID
      parameters:
      - description: Database Management ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of database objects
          schema:
            $ref: '#/definitions/controllers.DBObjectMgtListResponse'
        "400":
          description: Invalid database management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.DatabaseConnectionErrorResponse'
      summary: Get database objects by database management ID
      tags:
      - DB Object Management
...
  /api/queries/dbpolicy/dbmgt/{dbmgt}:
    get:
      consumes:
      - application/json
      description: Starts background job to generate database policies for specified
        database management ID
      parameters:
      - description: Database Management ID
        in: path
        name: dbmgt
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Background job started message with job ID
          schema:
            $ref: '#/definitions/controllers.JobStartResponse'
        "400":
          description: Invalid database management ID
          schema:
            $ref: '#/definitions/controllers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.JobProcessingErrorResponse'
      summary: Generate policies for database management
      tags:
      - DB Policy
...
swagger: "2.0"